requirements.txt
fastapi
uvicorn
pydantic<3
requests
python-dotenv
sentence-transformers
numpy
scikit-learn
streamlit
plotly
httpx
jinja2
openai>=1.0.0
anthropic>=0.8.0